refactor: change addBoardForm component
