+++ src/components/Menu/index.tsx
@@ -7,8 +7,7 @@ import cl from './menu.module.scss';
 import { useTranslation } from 'react-i18next';
 import { userSlice } from './../../store/reducers/userSlice';
 import { useAppDispatch, useAppSelector } from '../../hooks/redux';
-import BoardForm from '../BoardForm';
-import { addNewBoard, fetchAllBoards } from '../../store/reducers/actionCreators';
+import BoardForm from '../AddBoardForm';
 
 const Menu: React.FC = () => {
   const { user, lang } = useAppSelector((state) => state.user);
@@ -18,8 +17,6 @@ const Menu: React.FC = () => {
   const dispatch = useAppDispatch();
   const close = () => setOpen(false);
   const { setLang, logoutUser } = userSlice.actions;
-  const [title, setTitle] = useState<string>('');
-  const [description, setDescription] = useState<string>('');
   const [showModal, setShowModal] = useState(false);
 
   const logout = async () => {
@@ -39,11 +36,6 @@ const Menu: React.FC = () => {
     setShowModal(false);
   };
 
-  const createNewBoard = async () => {
-    await dispatch(addNewBoard({ title, description }));
-    await dispatch(fetchAllBoards());
-  };
-
   useEffect(() => {
     //throw new Error('Division by zero!');
     i18n.changeLanguage(lang);
@@ -55,10 +47,10 @@ const Menu: React.FC = () => {
         <nav className={`${cl.usermenu} ${open ? cl['usermenu-open'] : ''}`}>
           {!user.id && (
             <>
-              <Link onClick={() => close()} className={cl.usermenu_button} to="/signup">
+              <Link onClick={close} className={cl.usermenu_button} to="/signup">
                 {t('menu.signup')}
               </Link>
-              <Link onClick={() => close()} className={cl.usermenu_button} to="/signin">
+              <Link onClick={close} className={cl.usermenu_button} to="/signin">
                 {t('menu.login')}
               </Link>
             </>
@@ -68,14 +60,14 @@ const Menu: React.FC = () => {
               <Link className={cl.usermenu_button} to="/boards">
                 {t('menu.boards')}
               </Link>
-              <Link onClick={() => close()} className={cl.usermenu_button} to="/profile">
+              <Link onClick={close} className={cl.usermenu_button} to="/profile">
                 {t('menu.edit_profile')}
               </Link>
               <button onClick={handleOpenModal} className={cl.usermenu_button}>
                 {t('menu.new_board')}
               </button>
               <div className={cl.usermenu_username}>{user.login}</div>
-              <Link onClick={() => logout()} className={cl.usermenu_button} to="/">
+              <Link onClick={logout} className={cl.usermenu_button} to="/">
                 {t('menu.logout')}
               </Link>
             </>
@@ -90,16 +82,7 @@ const Menu: React.FC = () => {
         </nav>
         <Hamburger open={open} setOpen={setOpen} />
       </div>
-      <BoardForm
-        header={t('menu.new_board')}
-        title={title}
-        description={description}
-        setTitle={setTitle}
-        setDescription={setDescription}
-        onForm={createNewBoard}
-        onClose={handleClose}
-        show={showModal}
-      />
+      <BoardForm onClose={handleClose} show={showModal} />
     </>
   );
 };
